refactor(api): extract memory write operations to memory_writes.py
